+++ ft_memset.c
@@ -6,16 +6,17 @@
 /*   By: nguelfi <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/04/11 14:52:09 by nguelfi           #+#    #+#             */
-/*   Updated: 2017/04/11 15:33:27 by nguelfi          ###   ########.fr       */
+/*   Updated: 2017/04/13 17:51:12 by nguelfi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <string.h>
+#include "libft.h"
 
 void	*ft_memset(void *b, int c, size_t len)
 {
 	unsigned char	ch;
-	int				i;
+	size_t			i;
 	char			*dst;
 
 	i = 0;
